# Backend/src/main/resources/openapi/order.yaml
openapi: 3.0.0
info:
  title: Order Management API
  version: 1.0.0
  description: API for updating orders.
paths:
  /users/orders/{trackingid}:
    parameters:
      - in: path
        name: trackingid
        required: true
        schema:
          type: string
        description: The tracking ID of the order to be updated
    patch:
      tags:
        - orders
      summary: Update Order
      description: Update the details of a specific order.
      requestBody:
        description: Order status object to update.
        content:
          application/json:
            schema:
              type: object
              properties:
                order_status:
                  type: string
                  description: The new status of the order.
              required:
                - order_status
      responses:
        '200':
          description: Order updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_status:
                    type: string
                    description: The status of the order.
                  tracking_id:
                    type: string
                    description: The tracking ID of the order.
                  message:
                    type: string
                    description: Success message.
                    example: Order updated successfully.
        '404':
          description: Order not found.
        '400':
          description: Bad request.
...